+++ prix_banque_flutter_interface/lib/transfers_management/transfer_model.dart
@@ -1,8 +1,9 @@
+
 class Transfer {
   final int transferId;
   final String mailAdressTransferPayer;
   final String mailAdressTransferReceiver;
-  final int transferAmount;
+  final double transferAmount;
   final String transferType;
   final String receiverQuestion;
   final String receiverAnswer;
@@ -23,7 +24,7 @@ class Transfer {
       transferId: json['transferId'] as int,
       mailAdressTransferPayer: json['mailAdressTransferPayer'] as String,
       mailAdressTransferReceiver: json['mailAdressTransferReceiver'] as String,
-      transferAmount: json['transferAmount'] as int,
+      transferAmount: json['transferAmount'] as double,
       transferType: json['transferType'] as String,
       receiverQuestion: json['receiverQuestion'] as String,
       receiverAnswer: json['receiverAnswer'] as String,
